services/dewarp/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
python-multipart
torch
opencv-python
gdown
